libraryiot/requirements.txt
Django==3.0.8
gunicorn==20.0.4
django-heroku
djangorestframework==3.11.0
django-environ==0.4.5